Update main.py
I've moved in other scripts the initialization/objects and plot functions. I've tried to keep at least the variable's values in this script (as the coordinates of figures-retina) but i get a problem of "circular import dependency" as i was trying to import in the object_initialization script in the main_script and at the same time the values of the coordinates of the main_script in the object_initialization script.