Update anim_plot
i've added the button to pause/resume the animation with the counter of trials number. I think that it would be a good idea to have also like a cell on the plot were we can insert the numbers of the trials where to start and where to end the animation (i think that later we'll have a lot of trials and doing so we could choose the "block" of frames that we want to see in the animation); we can obviously do the same thing changing the code but anyway, maybe we can do it later - when we'll have more trials - because for now it could be a waste of time.