Add affordance and effect predictors
To train the system on affordance of different actions and effect of
them, I implemented an affordance predictor and an effect predictor
for each action.

The motivation signal needs to be changed to match these changes. The
same goes for data acquisition and graphics. For now, I made the
system always choose the parameterised skill, so the results are as
before these changes.