Change input of leaky average
This changes the input of the leaky average to the absolute value of
the change in affordance prediction, if an action is performed, and 0
otherwise. This is needed for the system to think changing its mind
about something based on new information is a good thing. The 0 input
will make the system interested in more things as time goes by and no
action is performed.