Reintroduce ignorance change as IM signal
Since the motivation signal for an object-action with target 0
decreases in the beginning when the second predictors weights are
initialized to give a small positive output, the linear ignorance is
reintroduced.

This measure creates another problem around the value 0.5 (at the tip
of the triangle), though. Keep this in mind!